Feat: Separate input container
